Fix and organize files
