--- src/components/appointments/NewAppointment.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-const NewAppointment = () => (
-  <div>
-    New Appointment
-  </div>
-);
-
-export default NewAppointment;
