--- src/components/doctors/AddDoctor.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-const NewDoctor = () => (
-  <div>
-    New Doctor
-  </div>
-);
-
-export default NewDoctor;
