--- src/components/doctors/DeleteDoctor.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-const DeleteDoctor = () => (
-  <div>
-    Delete Doctor
-  </div>
-);
-
-export default DeleteDoctor;
