--- src/components/doctors/DoctorDetails.js
@@ -1,8 +1,0 @@
-import React from 'react';
-
-const DoctorDetails = () => (
-  <div>
-    Doctor Details
-  </div>
-);
-export default DoctorDetails;
