--- src/components/doctors/DoctorsList.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-const DoctorsList = () => (
-  <div>
-    Doctors List
-  </div>
-);
-
-export default DoctorsList;
