--- src/components/specializations/SpecialisationsList.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-const SpecialisationsList = () => (
-  <div>
-    Specialisations List
-  </div>
-);
-
-export default SpecialisationsList;
