+++ src/routes/appointments/NewAppointment.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import SideNav from '../../components/home/SideNav';
+
+const NewAppointment = () => (
+  <div>
+
+    <SideNav />
+    <div>
+      New Appointment
+    </div>
+  </div>
+);
+
+export default NewAppointment;
