+++ src/routes/doctors/AddDoctor.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import SideNav from '../../components/home/SideNav';
+
+const NewDoctor = () => (
+  <div>
+    <SideNav />
+    <div>
+      New Doctor
+    </div>
+  </div>
+);
+
+export default NewDoctor;
