+++ src/routes/doctors/DeleteDoctor.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import SideNav from '../../components/home/SideNav';
+
+const DeleteDoctor = () => (
+  <div>
+    <SideNav />
+    <div>
+      Delete Doctor
+    </div>
+  </div>
+);
+
+export default DeleteDoctor;
