+++ src/routes/doctors/DoctorDetails.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import SideNav from '../../components/home/SideNav';
+
+const DoctorDetails = () => (
+  <div>
+    <SideNav />
+    <div>
+      Doctor details
+    </div>
+  </div>
+);
+export default DoctorDetails;
