+++ src/routes/doctors/slides.css
